+++ src/main/java/com/itransition/pmanager/dao/TagDao.java
@@ -7,6 +7,7 @@ import org.springframework.data.repository.query.Param;
 /**
  * Created by Lenovo on 16.06.2017.
  */
-public interface TagDao extends JpaRepository<Tag, Long> {
+public interface TagDao extends JpaRepository<Tag, Integer> {
     public Tag findByName(@Param("name") String name);
+    //public Set<Tag> findByProjects(@Param("projects")Project projects);
 }
